#45 Extract policy enforcement to policy.py
Move policy-related exception and enforcement functions out of ink_core.py into a new policy.py module (PolicyViolation, enforce_prompt_filter, enforce_deny_shell_commands, enforce_wrapper_policy, enforce_network_policy). Update ink_core.py to import these symbols, adjust install.py to copy policy.py into the private lib directory, and update tests/test_policy_enforcement.py to import the policy module and ensure ink_core is imported at the right time in the failing test. This refactors policy concerns into a dedicated module without changing enforcement behavior.
