+++ tests/test_policy_enforcement.py
@@ -51,7 +51,7 @@ import shutil
 import pytest
 from types import SimpleNamespace
 
-import ink_core as ink
+import policy as ink
 
 
 @pytest.fixture
@@ -371,6 +371,7 @@ def test_wrapper_passes_when_authenticated(monkeypatch, fake_state, fake_logging
 
 
 def test_main_short_circuits_before_subprocess(monkeypatch):
+    import ink_core as ink # Import here to ensure monkeypatching works before enforcement logic runs
     # Prevent actual subprocess execution
     monkeypatch.setattr(
         "ink_core.subprocess.run",
